--- apps/api-old/README.md
@@ -1,3 +1,0 @@
-# Dragonfish API
-
-The official Offprint API.
