--- apps/api-old/nest-cli.json
@@ -1,4 +1,0 @@
-{
-	"collection": "@nestjs/schematics",
-	"sourceRoot": "src"
-}
