--- apps/api-old/package.json
@@ -1,102 +1,0 @@
-{
-	"name": "@dragonfish/api",
-	"private": true,
-	"version": "1.0.0",
-	"description": "The official Offprint API",
-	"license": "MIT",
-	"scripts": {
-		"prebuild": "rimraf dist",
-		"build": "nest build",
-		"format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-		"start": "nest start",
-		"dev": "nest start --watch",
-		"start:debug": "nest start --debug --watch",
-		"start:prod": "node dist/main",
-		"lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
-		"test": "jest",
-		"test:watch": "jest --watch",
-		"test:cov": "jest --coverage",
-		"test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-		"test:e2e": "jest --config ./test/jest-e2e.json"
-	},
-	"dependencies": {
-		"@anchan828/nest-sendgrid": "^0.5.25",
-		"@nestjs/bull": "^0.4.2",
-		"@nestjs/common": "^8.2.4",
-		"@nestjs/core": "^8.2.4",
-		"@nestjs/event-emitter": "^1.0.0",
-		"@nestjs/graphql": "^9.1.2",
-		"@nestjs/jwt": "^8.0.0",
-		"@nestjs/mongoose": "^9.0.1",
-		"@nestjs/platform-express": "^8.2.4",
-		"@nestjsplus/cookies": "^1.1.1",
-		"@types/lodash": "^4.14.178",
-		"@types/mongoose-paginate-v2": "^1.4.1",
-		"@types/voca": "^1.4.2",
-		"apollo-server-core": "^3.5.0",
-		"argon2": "^0.28.3",
-		"body-parser": "^1.19.1",
-		"bull": "^4.1.4",
-		"cookie-parser": "^1.4.6",
-		"csurf": "^1.11.0",
-		"dotenv": "^10.0.0",
-		"graphql": "^16.1.0",
-		"helmet": "^4.6.0",
-		"lodash": "^4.17.21",
-		"mongoose": "^6.1.2",
-		"mongoose-autopopulate": "^0.16.0",
-		"mongoose-paginate-v2": "^1.4.2",
-		"mongoose-unique-validator": "^3.0.0",
-		"nanoid": "^3.1.30",
-		"reflect-metadata": "^0.1.13",
-		"rimraf": "^3.0.2",
-		"rxjs": "^7.4.0",
-		"sanitize-html": "^2.6.1",
-		"ua-parser-js": "^1.0.2",
-		"voca": "^1.4.0"
-	},
-	"devDependencies": {
-		"@dragonfish/models": "1.0.0",
-		"@nestjs/cli": "^8.1.6",
-		"@nestjs/schematics": "^8.0.5",
-		"@nestjs/testing": "^8.2.4",
-		"@types/bull": "^3.15.7",
-		"@types/cookie-parser": "^1.4.2",
-		"@types/csurf": "^1.11.2",
-		"@types/express": "^4.17.13",
-		"@types/jest": "^27.0.2",
-		"@types/node": "^16.11.1",
-		"@types/supertest": "^2.0.11",
-		"@typescript-eslint/eslint-plugin": "^4.29.2",
-		"@typescript-eslint/parser": "^4.29.2",
-		"eslint": "^7.32.0",
-		"eslint-config-prettier": "^8.3.0",
-		"eslint-plugin-prettier": "^3.4.1",
-		"jest": "^27.3.0",
-		"prettier": "^2.4.1",
-		"source-map-support": "^0.5.20",
-		"supertest": "^6.1.6",
-		"ts-jest": "^27.0.7",
-		"ts-loader": "^9.2.6",
-		"ts-node": "^10.3.0",
-		"tsconfig-paths": "^3.11.0",
-		"typescript": "^4.4.4"
-	},
-	"jest": {
-		"moduleFileExtensions": [
-			"js",
-			"json",
-			"ts"
-		],
-		"rootDir": "src",
-		"testRegex": ".*\\.spec\\.ts$",
-		"transform": {
-			"^.+\\.(t|j)s$": "ts-jest"
-		},
-		"collectCoverageFrom": [
-			"**/*.(t|j)s"
-		],
-		"coverageDirectory": "../coverage",
-		"testEnvironment": "node"
-	}
-}
