--- apps/api-old/tsconfig.json
@@ -1,23 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "module": "commonjs",
-        "declaration": true,
-        "removeComments": true,
-        "emitDecoratorMetadata": true,
-        "experimentalDecorators": true,
-        "allowSyntheticDefaultImports": true,
-        "target": "es2017",
-        "outDir": "./dist",
-        "baseUrl": "./",
-        "incremental": true,
-        "paths": {
-            "$shared/*": ["src/shared/*"],
-            "$db/*": ["src/db/*"],
-            "$modules/*": ["src/modules/*"]
-        }
-    },
-    "references": [
-        { "path": "../../packages/models" }
-    ]
-}
