+++ apps/api/src/modules/accounts/db/schemas/pseudonym.schema.ts
