+++ apps/api/src/shared/auth/index.ts
