+++ apps/api/src/shared/auth/optional.ts
