+++ apps/api/src/shared/auth/user.ts
