+++ apps/api/src/shared/util/device-info.ts
