+++ apps/api/src/shared/util/device.ts
