+++ apps/api/src/shared/util/sanitize-options.ts
