+++ apps/api/src/shared/util/secrets.ts
