+++ apps/api/tsconfig.json
@@ -9,8 +9,12 @@
     "target": "es2017",
     "sourceMap": true,
     "outDir": "./dist",
-    "baseUrl": "./",
+    "baseUrl": "./src",
     "incremental": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "paths": {
+        "$modules/*": ["src/modules/*"],
+        "$shared/*": ["src/shared/*"]
+    }
   }
 }
